requirements.txt
ag2==0.9.7
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.7
autogen-core==0.6.4
azure-common==1.1.28
azure-core==1.35.0
azure-search-documents==11.5.3
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.48.1
numpy==2.3.2
openai==1.97.1
opentelemetry-api==1.35.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
# pywin32==311
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.47.1
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zipp==3.23.0
google-cloud-storage>=2.14.0
